requirements.txt
streamlit
alpaca-py
pandas
numpy
ta
xgboost
scikit-learn
